centralize the excluded check
