+++ src/oblogger_view.ts
@@ -276,42 +276,18 @@ export class ObloggerView extends ItemView {
         this.registerInterval(this.renderTimeout);
     }
 
-    private getTemplatesFolders(): string[] {
-        const templatesFolders: string[] = []
-
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        const coreTemplatesFolder = this.app.internalPlugins.plugins["templates"]?.instance?.options?.folder ?? "";
-        if (coreTemplatesFolder !== "") {
-            templatesFolders.push(coreTemplatesFolder);
-        }
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        const templaterFolder = this.app.plugins.plugins["templater-obsidian"]?.settings?.templates_folder ?? "";
-        if (templaterFolder !== "") {
-            templatesFolders.push(templaterFolder);
-        }
-        return templatesFolders;
-    }
-
     private renderTagGroup(group: GroupFolder, modifiedFiles: FileState[]) {
         const maybeSettingsGroup = this.settings?.tagGroups.find(
             settingsGroup => settingsGroup.tag === group.groupName
         );
 
-        const collapsedFolders = maybeSettingsGroup?.collapsedFolders ?? [];
-        const excludedFolders = [...maybeSettingsGroup?.excludedFolders ?? []];
-        if (maybeSettingsGroup) {
-            if (!maybeSettingsGroup.templatesFolderVisible) {
-                excludedFolders.push(...this.getTemplatesFolders());
-            }
-            if (!maybeSettingsGroup.logsFolderVisible && this.settings) {
-                excludedFolders.push(this.settings.loggingPath);
-            }
+        if (!maybeSettingsGroup) {
+            console.warn(`unable to find settings for tag group ${group.groupName}`);
+            return;
         }
 
         if (group instanceof TagGroupContainer) {
-            group.render(collapsedFolders, excludedFolders, modifiedFiles, false);
+            group.render(modifiedFiles, false, maybeSettingsGroup);
         }
     }
 
@@ -325,23 +301,15 @@ export class ObloggerView extends ItemView {
             console.warn(`unable to find settings for rx group ${groupName}`);
             return;
         }
-        const excludedFolders = [...groupSetting.excludedFolders];
-        if (!groupSetting.templatesFolderVisible) {
-            excludedFolders.push(...this.getTemplatesFolders());
-        }
-        if (!groupSetting.logsFolderVisible && this.settings) {
-            excludedFolders.push(this.settings.loggingPath);
-        }
         const container = this.rxContainers.find(container => container.groupName === groupName);
         if (!container) {
             console.warn(`unable to find container for rx group ${groupName}`);
             return;
         }
         container.render(
-            groupSetting.collapsedFolders ?? [],
-            excludedFolders,
             modifiedFiles,
-            forced);
+            forced,
+            groupSetting);
     }
 
     private renderDailies(modifiedFiles: FileState[]) {
